Made changes in Stats.java class
Known problems:
In RaceCon.gamble if you enter a too high bet, then it repeats the race that many times as many times you had to repeat the betting sequence. The cause is known, it is due to a property of recursion. Planning to get it fixed soon, within a few days.